Handle null author image
Some servers return null for the image field on the author json. One
example is the Rust based
https://github.com/colinbankier/realworld-tide. If you create a new
profile with that server and this UI before these changes, you'll
get a "Deserialize Error". With this patch the author image field
changes from a String to an Option<String>, like bio.